RMO_Rate - rate changed to Double. HabitDetailVC - rate is created & % calculated when success button is pressed. MainVC - day difference error fixed. newHabitVC - rate creation removed
RMO_Rate - rate changed to Double to accurately show the %.
HabitDetailVC - RMO_Rate is no longer created when new Habit gets created. instead it gets created when success button is pressed. % is calculated when success button is pressed.
MainVC - day difference error fixed. It was adding by second using DayDifference variable. now it correctly adds "DAYS" using dayDifference var.
newHabitVC - rate creation removed. now RMO_Rate is created when success button is pressed.

+++ HabitBuilder/HabitBuilder/Database Model/RMO_Rate.swift
@@ -13,7 +13,7 @@ class RMO_Rate: Object {
     @Persisted var id = UUID().uuidString
     @Persisted var habitID: String = ""
     @Persisted var createdDate: Date = Date()
-    @Persisted var rate: Int
+    @Persisted var rate: Double
     
     //designating id as the primary key to prevent duplicate ones based on createdDate
     override static func primaryKey() -> String? {
